Initial port of keymap to 3x5 charybdis
Need to look into how the pointer layer is handled in the native version.
It seems link in the MCU rules.mk there is a `POINTING_DEVICE_DRIVER`
specified but I'm not sure what the layer looks like or if this same
behavior with the z and slsh key exist.